fix(e2e): align auth fixtures with onboarding and autosave
